+++ cmd/zetaclientd/keygen_tss.go
@@ -29,7 +29,9 @@ func GenerateTss(logger zerolog.Logger,
 	priKey secp256k1.PrivKey,
 	ts *metrics.TelemetryServer,
 	tssHistoricalList []observertypes.TSS,
-	metrics *metrics.Metrics) (*mc.TSS, error) {
+	metrics *metrics.Metrics,
+	tssPassword string,
+	hotkeyPassword string) (*mc.TSS, error) {
 	keygenLogger := logger.With().Str("module", "keygen").Logger()
 
 	// Bitcoin chain ID is currently used for using the correct signature format
@@ -49,6 +51,8 @@ func GenerateTss(logger zerolog.Logger,
 		tssHistoricalList,
 		metrics,
 		bitcoinChainID,
+		tssPassword,
+		hotkeyPassword,
 	)
 	if err != nil {
 		keygenLogger.Error().Err(err).Msg("NewTSS error")