+++ cmd/zetae2e/local/erc20.go
@@ -62,7 +62,7 @@ func erc20TestRoutine(
 		// run erc20 test
 		if err := erc20Runner.RunSmokeTestsFromNames(
 			smoketests.AllSmokeTests,
-			smoketests.TestWithdrawERC20Name,
+			smoketests.TestERC20WithdrawName,
 			smoketests.TestMultipleWithdrawsName,
 			smoketests.TestERC20DepositAndCallRefundName,
 			smoketests.TestZRC20SwapName,