+++ cmd/zetae2e/run.go
@@ -52,7 +52,7 @@ func runE2ETest(cmd *cobra.Command, args []string) error {
 	}
 
 	// initialize logger
-	logger := runner.NewLogger(verbose, color.FgWhite, "e2e")
+	logger := runner.NewLogger(verbose, color.FgHiCyan, "e2e")
 
 	// set config
 	app.SetConfig()
@@ -62,7 +62,7 @@ func runE2ETest(cmd *cobra.Command, args []string) error {
 
 	// get EVM address from config
 	evmAddr := conf.Accounts.EVMAddress
-	if ethcommon.IsHexAddress(evmAddr) {
+	if !ethcommon.IsHexAddress(evmAddr) {
 		cancel()
 		return errors.New("invalid EVM address")
 	}
@@ -75,8 +75,8 @@ func runE2ETest(cmd *cobra.Command, args []string) error {
 		conf,
 		ethcommon.HexToAddress(evmAddr),
 		conf.Accounts.EVMPrivKey,
-		utils.FungibleAdminName,
-		FungibleAdminMnemonic,
+		utils.FungibleAdminName, // placeholder value, not used
+		FungibleAdminMnemonic,   // placeholder value, not used
 		logger,
 	)
 	if err != nil {
@@ -88,18 +88,42 @@ func runE2ETest(cmd *cobra.Command, args []string) error {
 	logger.Print("starting tests")
 
 	// fetch the TSS address
-	testRunner.SetTSSAddresses()
+	if err := testRunner.SetTSSAddresses(); err != nil {
+		return err
+	}
 
-	// run tests
-	if err := testRunner.RunSmokeTestsFromNames(
-		smoketests.AllSmokeTests,
-		conf.TestList...,
-	); err != nil {
+	// set timeout
+	testRunner.CctxTimeout = 60 * time.Minute
+	testRunner.ReceiptTimeout = 60 * time.Minute
+
+	balancesBefore, err := testRunner.GetAccountBalances(true)
+	if err != nil {
 		cancel()
 		return err
 	}
 
+	//run tests
+	reports, err := testRunner.RunSmokeTestsFromNamesIntoReport(
+		smoketests.AllSmokeTests,
+		conf.TestList...,
+	)
+	if err != nil {
+		cancel()
+		return err
+	}
+
+	balancesAfter, err := testRunner.GetAccountBalances(true)
+	if err != nil {
+		cancel()
+		return err
+	}
+
+	// Print tests completion info
 	logger.Print("tests finished successfully in %s", time.Since(testStartTime).String())
+	testRunner.Logger.SetColor(color.FgHiRed)
+	testRunner.PrintTotalDiff(runner.GetAccountBalancesDiff(balancesBefore, balancesAfter))
+	testRunner.Logger.SetColor(color.FgHiGreen)
+	testRunner.PrintTestReports(reports)
 
 	return nil
 }