+++ contrib/localnet/orchestrator/smoketest/runner/bitcoin.go
@@ -27,6 +27,44 @@ import (
 
 var blockHeaderBTCTimeout = 5 * time.Minute
 
+// DepositBTCWithAmount deposits BTC on ZetaChain with a specific amount
+func (sm *SmokeTestRunner) DepositBTCWithAmount(amount float64) (txHash *chainhash.Hash) {
+	sm.Logger.Print("⏳ depositing BTC into ZEVM")
+
+	// fetch utxos
+	utxos, err := sm.BtcRPCClient.ListUnspentMinMaxAddresses(1, 9999999, []btcutil.Address{sm.BTCDeployerAddress})
+	if err != nil {
+		panic(err)
+	}
+
+	spendableAmount := 0.0
+	spendableUTXOs := 0
+	for _, utxo := range utxos {
+		if utxo.Spendable {
+			spendableAmount += utxo.Amount
+			spendableUTXOs++
+		}
+	}
+
+	if spendableAmount < amount {
+		panic(fmt.Errorf("not enough spendable BTC to run the test; have %f", spendableAmount))
+	}
+
+	sm.Logger.Info("ListUnspent:")
+	sm.Logger.Info("  spendableAmount: %f", spendableAmount)
+	sm.Logger.Info("  spendableUTXOs: %d", spendableUTXOs)
+	sm.Logger.Info("Now sending two txs to TSS address...")
+
+	amount = amount + zc_btc.BtcDepositorFeeMin
+	txHash, err = sm.SendToTSSFromDeployerToDeposit(sm.BTCTSSAddress, amount, utxos, sm.BtcRPCClient, sm.BTCDeployerAddress)
+	if err != nil {
+		panic(err)
+	}
+	sm.Logger.Info("send BTC to TSS txHash: %s", txHash.String())
+
+	return txHash
+}
+
 // DepositBTC deposits BTC on ZetaChain
 func (sm *SmokeTestRunner) DepositBTC(testHeader bool) {
 	sm.Logger.Print("⏳ depositing BTC into ZEVM")
@@ -128,7 +166,7 @@ func (sm *SmokeTestRunner) SendToTSSFromDeployerWithMemo(
 	to btcutil.Address,
 	amount float64,
 	inputUTXOs []btcjson.ListUnspentResult,
-	btc *rpcclient.Client,
+	btcRPC *rpcclient.Client,
 	memo []byte,
 	btcDeployerAddress *btcutil.AddressWitnessPubKeyHash,
 ) (*chainhash.Hash, error) {
@@ -157,8 +195,9 @@ func (sm *SmokeTestRunner) SendToTSSFromDeployerWithMemo(
 		btcDeployerAddress: change,
 	}
 
-	// create raw transaction
-	tx, err := btc.CreateRawTransaction(inputs, amountMap, nil)
+	// create raw
+	sm.Logger.Info("ADDRESS: %s, %s", btcDeployerAddress.EncodeAddress(), to.EncodeAddress())
+	tx, err := btcRPC.CreateRawTransaction(inputs, amountMap, nil)
 	if err != nil {
 		panic(err)
 	}
@@ -201,38 +240,42 @@ func (sm *SmokeTestRunner) SendToTSSFromDeployerWithMemo(
 		}
 	}
 
-	stx, signed, err := btc.SignRawTransactionWithWallet2(tx, inputsForSign)
+	stx, signed, err := btcRPC.SignRawTransactionWithWallet2(tx, inputsForSign)
 	if err != nil {
 		panic(err)
 	}
 	if !signed {
 		panic("btc transaction not signed")
 	}
-	txid, err := btc.SendRawTransaction(stx, true)
+	txid, err := btcRPC.SendRawTransaction(stx, true)
 	if err != nil {
 		panic(err)
 	}
 	sm.Logger.Info("txid: %+v", txid)
-	_, err = btc.GenerateToAddress(6, btcDeployerAddress, nil)
+	_, err = btcRPC.GenerateToAddress(6, btcDeployerAddress, nil)
 	if err != nil {
 		panic(err)
 	}
-	gtx, err := btc.GetTransaction(txid)
+	gtx, err := btcRPC.GetTransaction(txid)
 	if err != nil {
 		panic(err)
 	}
 	sm.Logger.Info("rawtx confirmation: %d", gtx.BlockIndex)
-	rawtx, err := btc.GetRawTransactionVerbose(txid)
+	rawtx, err := btcRPC.GetRawTransactionVerbose(txid)
 	if err != nil {
 		panic(err)
 	}
 
+	btcChainID, err := common.GetBTCChainIDFromChainParams(sm.BitcoinParams)
+	if err != nil {
+		panic(err)
+	}
 	events := zc_btc.FilterAndParseIncomingTx(
 		[]btcjson.TxRawResult{*rawtx},
 		0,
 		sm.BTCTSSAddress.EncodeAddress(),
 		&log.Logger,
-		common.BtcRegtestChain().ChainId,
+		btcChainID,
 	)
 	sm.Logger.Info("bitcoin intx events:")
 	for _, event := range events {